lesson 7: make исправление недочетов 2
