--- TodoList/Properties/launchSettings.json
@@ -1,1 +1,0 @@
-{}
